Create `pkg/driver/version` package to provide version information
